langgraph_orchestrator/requirements.txt
# LangGraph Core
langgraph>=0.5.0
langchain>=0.1.0
langchain-core>=0.1.0

# AI Models
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP & Async
httpx>=0.25.0
aiohttp>=3.8.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional Dependencies
redis>=5.0.0
faiss-cpu>=1.7.4

# Development
pytest>=7.0.0
black>=23.0.0 